Edited edit_event.html file
Hi there 

I'm samiullah and for the bug I have brought a solution to fix it which involves creating an aestrik on the form elements. I have also looked into the demo json file and tried to adjust it to that. The main issue of fixing the missing form indicators has been fixed and the above code illustrates this. 

Can you approve of my solution please.